refactor: Move ConversationAccount and DraftAttachment to core.model
Needs a JSON adapter for URIs, so write one.
