+++ core/model/src/main/kotlin/app/pachli/core/model/DraftAttachment.kt
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2025 Pachli Association
+ *
+ * This file is a part of Pachli.
+ *
+ * This program is free software; you can redistribute it and/or modify it under the terms of the
+ * GNU General Public License as published by the Free Software Foundation; either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * Pachli is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
+ * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+ * Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with Pachli; if not,
+ * see <http://www.gnu.org/licenses>.
+ */
+
+package app.pachli.core.model
+
+import android.net.Uri
+import android.os.Parcelable
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+@JsonClass(generateAdapter = true)
+data class DraftAttachment(
+    @Json(name = "uriString") val uri: Uri,
+    val description: String?,
+    val focus: Attachment.Focus?,
+    val type: Type,
+) : Parcelable {
+    enum class Type {
+        IMAGE,
+        VIDEO,
+        AUDIO,
+    }
+}
