+++ core/network/src/main/kotlin/app/pachli/core/network/di/NetworkModule.kt
@@ -29,6 +29,7 @@ import app.pachli.core.network.json.Guarded
 import app.pachli.core.network.json.HasDefault
 import app.pachli.core.network.json.InstantJsonAdapter
 import app.pachli.core.network.json.LenientRfc3339DateJsonAdapter
+import app.pachli.core.network.json.UriAdapter
 import app.pachli.core.network.model.MediaUploadApi
 import app.pachli.core.network.retrofit.InstanceSwitchAuthInterceptor
 import app.pachli.core.network.retrofit.MastodonApi
@@ -68,6 +69,7 @@ object NetworkModule {
     fun providesMoshi(): Moshi = Moshi.Builder()
         .add(Date::class.java, LenientRfc3339DateJsonAdapter())
         .add(Instant::class.java, InstantJsonAdapter())
+        .add(UriAdapter())
         .add(VersionAdapter())
         .add(Guarded.Factory())
         .add(HasDefault.Factory())
