+++ core/network/src/main/kotlin/app/pachli/core/network/json/UriAdapter.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2025 Pachli Association
+ *
+ * This file is a part of Pachli.
+ *
+ * This program is free software; you can redistribute it and/or modify it under the terms of the
+ * GNU General Public License as published by the Free Software Foundation; either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * Pachli is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
+ * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+ * Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with Pachli; if not,
+ * see <http://www.gnu.org/licenses>.
+ */
+
+package app.pachli.core.network.json
+
+import android.net.Uri
+import androidx.core.net.toUri
+import com.squareup.moshi.FromJson
+import com.squareup.moshi.ToJson
+
+class UriAdapter {
+    @ToJson
+    fun toJson(uri: Uri): String = uri.toString()
+
+    @FromJson
+    fun fromJson(s: String): Uri = s.toUri()
+}
